fix: display of the eloRating is correct

+++ src/main/resources/db/migration/V01__create_player_table.sql
@@ -2,6 +2,6 @@ CREATE TABLE player
 (
     player_id     BIGINT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
     nickname      TEXT      NOT NULL UNIQUE,
-    elo_rating    INTEGER            DEFAULT 1200 NOT NULL,
+    elo_rating    NUMERIC(10, 2)     DEFAULT 1200.00 NOT NULL,
     registered_at TIMESTAMP NOT NULL DEFAULT now()
 );
